adição das views de Dentista adição do CRUD de dentista em DentistaController.php tradução das mensagens de validação do back-end

+++ database/migrations/2022_03_28_134424_create_dentistas_table.php
@@ -17,7 +17,7 @@ class CreateDentistasTable extends Migration
             $table->bigIncrements('id')->nullable(false);
             $table->string('name', 100)->nullable(false);
             $table->string('email', 100)->nullable(false);
-            $table->integer('cro')->nullable(false);
+            $table->bigInteger('cro')->nullable(false);
             $table->char('cro_uf', 2)->nullable(false);
             $table->timestamps();
             $table->softDeletes();
